advances useries plots
